+++ src/main/java/org/example/Building/Enclosure/BirdEnclosure.java
@@ -1,6 +1,6 @@
-package org.example.Buildings.Enclosure;
+package org.example.Building.Enclosure;
 
-import org.example.Buildings.Enclosure;
+import org.example.Building.Enclosure;
 
 public class BirdEnclosure extends Enclosure {
