+++ src/main/java/org/example/Building/Shop/Drinks.java
@@ -1,0 +1,4 @@
+package org.example.Building.Shop;
+
+public class Drinks {
+}
