+++ src/main/java/org/example/Building/Shop/Tickets.java
@@ -1,0 +1,4 @@
+package org.example.Building.Shop;
+
+public class Tickets {
+}
