+++ src/main/java/org/example/Buildings.java
@@ -1,4 +1,9 @@
 package org.example;
 
-public abstract class Buildings {
+public abstract class Building{
+    protected String address;
+
+    public Building(String address){
+        this.address = address;
+    }
 }
