--- src/main/java/org/example/Buildings/Shop/Drinks.java
@@ -1,4 +1,0 @@
-package org.example.Buildings.Shop;
-
-public class Drinks {
-}
