--- src/main/java/org/example/Buildings/Shop/Food.java
@@ -1,4 +1,0 @@
-package org.example.Buildings.Shop;
-
-public class Food {
-}
