--- src/main/java/org/example/Buildings/Shop/Gifts.java
@@ -1,5 +1,0 @@
-package org.example.Buildings.Shop;
-
-public class Gifts {
-
-}
