--- src/main/java/org/example/Buildings/Shop/Tickets.java
@@ -1,4 +1,0 @@
-package org.example.Buildings.Shop;
-
-public class Tickets {
-}
